Swap go's net/http ServeMux with gorilla/mux Router
This allows is to leave paths alone, allowing for double-slashes to
pass through the proxy. For example, previously a path like
".../%2Fchanges/..." would result in a 301-redirect to
".../changes/...". This resulted in a bad request to the upstream
service. With this patch, the encoded slash is kept and the upstream
service is happy.

Bump version to v2.2.2.